Add files via upload
"Netflix Homepage Replica"

This GitHub repository contains a recreation of Netflix's homepage. Dive into the world of streaming entertainment with this project, which aims to replicate the look and feel of the Netflix homepage. Explore the code and design elements used to create this user interface and learn about web development techniques to build your own attractive and functional website.